Add exact_match function to see if two molecules are the same
In order to compare if two molecules are the same, it is insufficient to
just compare the SMILES because a molecular graph can be traversed
differently and return different SMILES for the same molecule, if I have
understood things correctly. InCHI can also have some issues.

There also are issues with tautomers and chirality

See here for more discussion [1]

The code to compare two molecules is ported from the code in
chemicalite and the rdkit postgres extension, many thanks for their work.

I tried to follow the rdkit postgres extension pattern -- having an
exact match operator '@=' but haven't figured out how to do that in
duckdb. So for now, there is an exact_match function

[1]: https://sourceforge.net/p/rdkit/mailman/rdkit-discuss/thread/AM0PR04MB5315B6D8DE8B380CAAF1211FE2AF9@AM0PR04MB5315.eurprd04.prod.outlook.com/