Fix memory leak from RDKit::ROMol missuse
There seems to be a particular way to use RDKit::ROMol when converting
molecules from smiles to mol to pickle and back. I tested things out in
a simple program that just converts a smiles to mol, then pickles it,
and then deserializes it back to a mol.

The way that seems to work, also the way it is done in chemicalite, is
to use a unique_ptr and to reset that with the smiles. This was one
memory leak. Previously I just used a regular pointer.

When deserializing the pickled molecule, it seems important to create a
new RDKit::ROMol in the unique_ptr. I think this creates it on the heap
as opposed to the stack. I don't know if that's correct, but I also
don't know why the object is not destroyed when it goes out of scope, as
I had expected from a unique_ptr. But I am a novice to C++ so perhaps I
am just misunderstanding all of this.

Anyhow, taking care of these two things solves the memory leak. I ran
duckdb with valgrind (`valgrind ./build/releases/duckdb`) and ran
through some of the conversion functions and it reports all blocks are
freed